Consolidate style into one styles file
